chest-xray-analyzer/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pillow==10.2.0
pydicom>=2.4.0
pylibjpeg>=1.4.0
torch>=2.2.0
torchvision>=0.17.0
torchxrayvision==1.2.1
numpy>=1.24.0
opencv-python>=4.9.0
scikit-image>=0.22.0